recipe-scraper-backend/requirements.txt
Flask==2.1.1
flask-restx==0.5.1
requests==2.28.1
beautifulsoup4==4.10.0
Flask-Cors==3.0.10
PyJWT==2.3.0